# src/zenml/zen_server/openapi3_0.yaml
openapi: 3.0.2
info:
  title: ZenML
  version: 0.xx.0
  description: >-
    The Zen API allows access to ZenML programmatically. These docs DO NOT represent the status quo but instead are the Design that is being aimed at.
...
paths:
...
  /pipelines/{pipeline_id}/runs:
    get:
      summary: Get pipeline runs of a specific pipeline
      tags:
        - pipeline
      description: |-
        Returns a list of runs from the pipeline.
      operationId: get_pipeline_runs_in_pipeline
      parameters:
        - $ref: '#/components/parameters/pipelineId'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List the pipeline runs of the pipeline
                type: array
                items:
                  $ref: '#/components/schemas/Run'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    post:
      summary: Create a new run for a pipeline
      tags:
        - pipeline
      description: |-
        Create a run for a pipeline. This endpoint is not meant to be used explicitly once ZenML follows the 
        centralized paradigm where runs are authored by the ZenServer and not on the users machine.
      operationId: create_pipeline_run_for_pipeline
      parameters:
        - $ref: '#/components/parameters/pipelineId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Run'
        required: true
      responses:
        '200': 
          $ref: '#/components/responses/200'
        '401': 
          $ref: '#/components/responses/401'
        '409': 
          $ref: '#/components/responses/409'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
...
  /runs:
    get:
      summary: Get all pipeline runs
      tags:
        - run
      description: |-
        Get pipeline runs according to query filters.
      operationId: get_pipeline_runs
      parameters:    
        - in: query
          name: project_name
          required: false
          schema:
            title: Project Name
            type: string
        - name: stack_id
          in: query
          required: false
          schema:
            title: Stack Id
            type: string
        - name: pipeline_id
          in: query
          required: false
          schema:
            title: Pipeline Id
            type: string
        - name: trigger_id
          in: query
          required: false
          schema:
            title: Trigger Id
            type: string
      responses:
        '200':
          description: Returns the requested pipeline run.
          content:
            application/json:
              schema:
                title: List all pipeline runs
                type: array
                items:
                  $ref: '#/components/schemas/Run'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'

      security:
        - HTTPBasic: []
  /runs/{run_id}:
    get:
      summary: Get a specific pipeline run
      tags:
        - run
      description: |-
        Get a specific pipeline run using its unique id.
      operationId: get_pipeline_run_by_id
      parameters:
        - $ref: '#/components/parameters/runId'
      responses:
        '200':
          description: Returns the requested pipeline run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'

      security:
        - HTTPBasic: []
    put:
      summary: Update a specific pipeline run
      tags:
        - run
      description: |-
        Updates the attributes on a pipeline run.
      operationId: update_pipeline_run_by_id
      parameters:
        - $ref: '#/components/parameters/runId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Run'
        required: true
      responses:
        '200':
          description: Returns the updated pipeline run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    delete:
      summary: Delete a specific pipeline run
      tags:
        - run
      description: |-
        Deletes a pipeline run
      operationId: delete_pipeline_run_by_id
      parameters:
        - $ref: '#/components/parameters/runId'
      responses:
        '200': 
          $ref: '#/components/responses/200'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
...
  /users/{user_id}/roles:
    get:
      summary: Get all roles assigned to a user
      tags:
        - user
      description: |-
        Returns all roles that are assigned to a user
      operationId: get_roles_for_user
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List of Roles assigned to user
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    post:
      summary: Assign a role to a user
      tags:
        - user
      description: |-
        Assign a role to a user for all resources within a given project or globally.
      operationId: assign_role_to_user
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
        required: true
      responses:
        '200': 
          $ref: '#/components/responses/200'
        '401': 
          $ref: '#/components/responses/401'
        '409': 
          $ref: '#/components/responses/409'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
...
  /users/{user_id}/roles/{role_id}:
    delete:
      summary: Remove a users role.
      tags:
        - user
      description: |-
        Remove a users role within a project or globally.
      operationId: remove_user_role_in_project
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/roleId'
        - in: query
          required: false
          schema:
            title: Project Name
            type: string
          name: project_name
      responses:
        '200': 
          $ref: '#/components/responses/200'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
  #-------------------------------
  # Roles
  #-------------------------------
  /roles:
    get:
      summary: Roles
      tags:
        - role
      description: |-
        Returns a list of all roles.
      operationId: get_roles_list      
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List of Roles
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
  #-------------------------------
  # Auth  related things
  #-------------------------------
...
components:
...
  schemas:
...
    Run:
      title: Run
      required:
        - id
        - name
        - pipeline
        - stack
        - runtime_configuration
        - user_id
        - project_name
      type: object
      properties:
        id:
          title: id
          type: string
          readOnly: True
          example: 68be3034
        name:
          title: Name
          type: string
          example: cat_detection_training_pipeline_run_4921
        zenml_version:
          title: Zenml Version
          type: string
          description: >-
            The ZenML Version that the pipeline was run on.
          example: 0.12.0
        duration:
          title: duration
          type: string
          description: >-
            The dureation of the last run.
          readOnly: True
        git_sha:
          title: Git Sha
          type: string
          description: >-
            The GitSha of the code repository this pipeline was run with
          example: 73841b4
        pipeline:
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string   
          description: >-
            The pipeline that the run is an instance of.
          example:
            {name: cat_detection_training_pipeline, id: 9eaa2d6}
        stack:
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string
          description: >-
            The stack that was used for this run.
          example:
            {name: prod_stack, id: 9lg2jk3}
        trigger:
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string
          description: >-
            The trigger that was used for this run.
          example:
            {name: weekly_prod_trigger, id: 381jd1k}
        status:
          type: string
          enum:
            - Running
            - Failed
            - Finished
          readOnly: True
          example: Finished
        owner:
          title: Owner
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string   
          readOnly: True
          description: >-
            User that owns this run.
          example:
            {name: htahir, id: w813dw}
        project_name:
          title: Project Name
          type: string
          example: demo_prj
        created_at:
          title: Run started at 
          type: string
          format: date-time
          readOnly: True
          description: >-
            Run started at.
          example: 2022-08-12T07:12:45.931Z
      description: |-
        A single Pipeline run. 
...
    Role:
      title: Role
      required:
        - name
      type: object
      properties:
        id:
          title: Id
          type: string
          format: uuid
          readOnly: True
        created_at:
          title: Creation Date
          type: string
          format: date-time
          readOnly: True
        name:
          title: Name
          type: string
      description: |-
        Roles that a user can have in a global/project setting.
    ErrorModel:
      title: ErrorModel
      type: object
      properties:
        detail:
          title: Detail
      description: Base class for error responses.
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
  #-------------------------------
  # Reusable operation parameters
  #-------------------------------
  parameters:
    pipelineId:
      name: pipeline_id
      in: path
      required: true
      schema:
        title: Pipeline ID
        type: string
...
    runId:
      name: run_id
      in: path
      required: true
      schema:
        title: Run ID
        type: string
...
    userId:
      name: user_id
      in: path
      required: true
      schema:
        title: User ID
        type: string
    roleId:
      name: role_id
      in: path
      required: true
      schema:
        title: Role Id
        type: string
...
  responses:
    '200':
      description: Successful Response
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'        
    '409':
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    '422':
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
...
  securitySchemes:
    HTTPBasic:
      type: http
      scheme: basic